refactor: Migrate the queue tables to the shared table state hook
PR 3 of the #366 cut, resolving #395 — the first relocation, done on the
simplest pair. useReactTable moves out of WantedTable and UpcomingTable
into their pages, which already owned selection because they render
BulkActionBar as a sibling; now they own the table instance and read
selectedIds off the hook. The components become dumb renderers taking
`table` (DataTable already does), with the shared columns extracted to
wantedColumns.tsx / upcomingColumns.tsx so the pages and the selection
tests drive the same select column.

- WantedPage adopts useServerPage(50); UpcomingTable is unpaginated, so
  MyReleasesView simply omits `pagination` (#360, #366's correction of
  #360's count).
- The hand-rolled pruning effects (WantedTable:57, UpcomingTable:48) are
  deleted — load-bearing today (TanStack never prunes, by design, #355),
  now owned by the hook. Retired *into* the hook, not dropped.
- selectAllScope: "filtered" on both. Neither imported getFilteredRowModel,
  getPaginationRowModel or globalFilter, so getIsAllRowsSelected() and
  getIsAllPageRowsSelected() are provably identical here — the parameter
  is semantics, changing behaviour at no call site (#382).
- getRowId: (row) => row.IssueID on both — entity ids, audited sound
  (#364). No identity fix here.
- Parity: all seven QueueTableSelection pins were table-state pins and
  none survived verbatim. The five generic #307 pins now live in
  useTableState.test.tsx carrying their #307 reference (three landed with
  #401; header select-all and single-row deselect land here). The two
  decode pins stay per-site, driven through harnesses that stand in for
  the pages as they now are: calling useTableState and passing `table`
  down. No pin retired before its replacement landed.
- Both files leave the ESLint ratchet allowlist.

Changeset: none — relocation only; both tables already pruned, and the
clearSelection leak is latent (#365). Nothing an operator can observe.

Suite: 29 files / 171 tests green; typecheck and every lint lane clean.
