+++ frontend/tests/components/QueueTableSelection.test.tsx
@@ -1,9 +1,12 @@
-import { useState } from "react";
+import { useEffect } from "react";
 import { describe, expect, it, vi } from "vitest";
 import userEvent from "@testing-library/user-event";
 import { render, screen } from "../test-utils";
 import WantedTable from "@/components/queue/WantedTable";
+import { useWantedColumns } from "@/components/queue/wantedColumns";
 import UpcomingTable from "@/components/queue/UpcomingTable";
+import { useUpcomingColumns } from "@/components/queue/upcomingColumns";
+import { useTableState } from "@/components/data-table/useTableState";
 import type { Issue } from "@/types";
 
 /**
@@ -12,9 +15,16 @@ import type { Issue } from "@/types";
  * was filtered out, and the bulk action bar reported nothing selected -- so
  * every bulk action silently ran on zero issues.
  *
- * Selection is now owned by the page and passed back down, so these tests drive
- * the tables the way the pages do: a controlled harness that echoes each
- * reported selection back as the `selectedIds` prop.
+ * These are the two per-site pins of that decode: each table's real select
+ * column, driven through its real DOM, must report entity ids. The generic
+ * halves of the old suite -- non-adjacent multi-select, header select-all,
+ * deselect, not resurrecting cleared ids, dropping ids whose rows left --
+ * moved to useTableState.test.tsx with their #307 reference when #395 moved
+ * the table instance into the pages.
+ *
+ * The harnesses stand in for WantedPage / MyReleasesView as they now are:
+ * calling useTableState with the shared columns, passing `table` down, and
+ * reading `selectedIds` off the hook.
  */
 
 function issues(count: number): Issue[] {
@@ -32,42 +42,53 @@ function issues(count: number): Issue[] {
   });
 }
 
-type QueueTable = typeof WantedTable | typeof UpcomingTable;
+function useReportedSelection(
+  selectedIds: string[],
+  onSelectionChange: (ids: string[]) => void,
+) {
+  useEffect(() => {
+    if (selectedIds.length > 0) onSelectionChange(selectedIds);
+  }, [selectedIds, onSelectionChange]);
+}
 
-/**
- * Stands in for WantedPage / MyReleasesView: holds the selection, feeds it back
- * to the table, and exposes the page-side Clear button plus a data swap.
- */
-function SelectionHarness({
-  table: Table,
+function WantedHarness({
   rows,
   onSelectionChange,
 }: {
-  table: QueueTable;
   rows: Issue[];
   onSelectionChange: (ids: string[]) => void;
 }) {
-  const [selectedIds, setSelectedIds] = useState<string[]>([]);
-  const [data, setData] = useState(rows);
+  const columns = useWantedColumns();
+  const { table, selectedIds } = useTableState({
+    data: rows,
+    columns,
+    getRowId: (row) => row.IssueID,
+    selection: { scope: "filtered" },
+    initialSorting: [{ id: "DateAdded", desc: true }],
+  });
+  useReportedSelection(selectedIds, onSelectionChange);
 
-  return (
-    <div>
-      <button type="button" onClick={() => setSelectedIds([])}>
-        Clear Selection
-      </button>
-      <button type="button" onClick={() => setData(data.slice(0, 1))}>
-        Drop rows
-      </button>
-      <Table
-        issues={data}
-        selectedIds={selectedIds}
-        onSelectionChange={(ids: string[]) => {
-          setSelectedIds(ids);
-          onSelectionChange(ids);
-        }}
-      />
-    </div>
-  );
+  return <WantedTable table={table} />;
+}
+
+function UpcomingHarness({
+  rows,
+  onSelectionChange,
+}: {
+  rows: Issue[];
+  onSelectionChange: (ids: string[]) => void;
+}) {
+  const columns = useUpcomingColumns();
+  const { table, selectedIds } = useTableState({
+    data: rows,
+    columns,
+    getRowId: (row) => row.IssueID,
+    selection: { scope: "filtered" },
+    initialSorting: [{ id: "IssueDate", desc: false }],
+  });
+  useReportedSelection(selectedIds, onSelectionChange);
+
+  return <UpcomingTable table={table} />;
 }
 
 function rowCheckboxes(): HTMLElement[] {
@@ -84,11 +105,7 @@ describe("queue table selection", () => {
     const user = userEvent.setup();
     const onSelectionChange = vi.fn();
     render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(3)}
-        onSelectionChange={onSelectionChange}
-      />,
+      <WantedHarness rows={issues(3)} onSelectionChange={onSelectionChange} />,
     );
 
     await user.click(rowCheckboxes()[0]);
@@ -101,8 +118,7 @@ describe("queue table selection", () => {
     const user = userEvent.setup();
     const onSelectionChange = vi.fn();
     render(
-      <SelectionHarness
-        table={UpcomingTable}
+      <UpcomingHarness
         rows={issues(3)}
         onSelectionChange={onSelectionChange}
       />,
@@ -111,114 +127,6 @@ describe("queue table selection", () => {
     await user.click(rowCheckboxes()[0]);
 
     expect(onSelectionChange).toHaveBeenCalled();
-    expect(lastSelection(onSelectionChange)).toEqual(["issue-1"]);
-  });
-
-  it("reports ids, not row positions, for a non-adjacent multi-row selection", async () => {
-    const user = userEvent.setup();
-    const onSelectionChange = vi.fn();
-    render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(4)}
-        onSelectionChange={onSelectionChange}
-      />,
-    );
-
-    await user.click(rowCheckboxes()[1]);
-    await user.click(rowCheckboxes()[3]);
-
-    // A position-based decode would report issue-1/issue-2 here.
-    expect(lastSelection(onSelectionChange)?.sort()).toEqual([
-      "issue-2",
-      "issue-4",
-    ]);
-  });
-
-  it("reports every id for the header select-all toggle", async () => {
-    const user = userEvent.setup();
-    const onSelectionChange = vi.fn();
-    render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(3)}
-        onSelectionChange={onSelectionChange}
-      />,
-    );
-
-    await user.click(screen.getAllByRole("checkbox")[0]);
-
-    expect(lastSelection(onSelectionChange)?.sort()).toEqual([
-      "issue-1",
-      "issue-2",
-      "issue-3",
-    ]);
-  });
-
-  it("drops an issue from the selection when it is unchecked", async () => {
-    const user = userEvent.setup();
-    const onSelectionChange = vi.fn();
-    render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(3)}
-        onSelectionChange={onSelectionChange}
-      />,
-    );
-
-    await user.click(rowCheckboxes()[0]);
-    await user.click(rowCheckboxes()[0]);
-
-    expect(lastSelection(onSelectionChange)).toEqual([]);
-  });
-
-  it("does not resurrect ids the page cleared", async () => {
-    const user = userEvent.setup();
-    const onSelectionChange = vi.fn();
-    render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(4)}
-        onSelectionChange={onSelectionChange}
-      />,
-    );
-
-    await user.click(rowCheckboxes()[0]);
-    await user.click(rowCheckboxes()[1]);
-    expect(lastSelection(onSelectionChange)?.sort()).toEqual([
-      "issue-1",
-      "issue-2",
-    ]);
-
-    await user.click(screen.getByRole("button", { name: "Clear Selection" }));
-    await user.click(rowCheckboxes()[3]);
-
-    // Before selection was lifted to the page, the table kept its own copy and
-    // reported issue-1 and issue-2 again -- skipping issues the user cleared.
-    expect(lastSelection(onSelectionChange)).toEqual(["issue-4"]);
-  });
-
-  it("drops selected ids whose rows are no longer rendered", async () => {
-    const user = userEvent.setup();
-    const onSelectionChange = vi.fn();
-    render(
-      <SelectionHarness
-        table={WantedTable}
-        rows={issues(3)}
-        onSelectionChange={onSelectionChange}
-      />,
-    );
-
-    await user.click(rowCheckboxes()[0]);
-    await user.click(rowCheckboxes()[2]);
-    expect(lastSelection(onSelectionChange)?.sort()).toEqual([
-      "issue-1",
-      "issue-3",
-    ]);
-
-    await user.click(screen.getByRole("button", { name: "Drop rows" }));
-
-    // issue-3 is gone from the data, so a bulk action must not still target it.
     expect(lastSelection(onSelectionChange)).toEqual(["issue-1"]);
   });
 });
